requirements.txt
Django==3.2.13
djangorestframework==3.13.1
django-filter==2.4
psycopg2-binary==2.9.3
gunicorn==20.1.0
